+++ deployment/terraform/assets/provisioners/rhel/openldap.sh
@@ -36,74 +36,6 @@ sudo chmod 700 /var/lib/ldap/
 sudo systemctl start slapd
 sudo systemctl enable slapd
 
-# Create base configuration LDIF
-cat << EOF > /tmp/base_config.ldif
-dn: olcDatabase={2}mdb,cn=config
-changetype: modify
-replace: olcSuffix
-olcSuffix: ${LDAP_BASE_DN}
-
-dn: olcDatabase={2}mdb,cn=config
-changetype: modify
-replace: olcRootDN
-olcRootDN: ${LDAP_ADMIN_DN}
-
-dn: olcDatabase={2}mdb,cn=config
-changetype: modify
-replace: olcRootPW
-olcRootPW: ${LDAP_ADMIN_PASSWORD_HASH}
-EOF
-
-# Apply base configuration
-sudo ldapmodify -Y EXTERNAL -H ldapi:/// -f /tmp/base_config.ldif
-
-# Create base DN structure
-cat << EOF > /tmp/base_structure.ldif
-dn: ${LDAP_BASE_DN}
-objectClass: top
-objectClass: dcObject
-objectClass: organization
-o: ${LDAP_ORGANIZATION}
-dc: mm
-
-dn: ou=users,${LDAP_BASE_DN}
-objectClass: organizationalUnit
-ou: users
-
-dn: ou=groups,${LDAP_BASE_DN}
-objectClass: organizationalUnit
-ou: groups
-description: Container for group accounts
-EOF
-
-# Add base structure
-sudo ldapadd -x -D "${LDAP_ADMIN_DN}" -w "${LDAP_ADMIN_PASSWORD}" -f /tmp/base_structure.ldif
-
-# Create sample LDIF file for test users
-cat << EOF > /tmp/test_users.ldif
-# Mattermost Users LDIF Export
-# Generated on Mon Jun 23 08:02:05 PM IST 2025
-
-dn: uid=testuser-1,ou=users,${LDAP_BASE_DN}
-objectClass: inetOrgPerson
-objectClass: person
-objectClass: top
-uid: testuser-1
-cn: testuser-1
-sn: User
-mail: testuser-1@mattermost.com
-userPassword: testPass123$
-
-dn: cn=developers,ou=groups,${LDAP_BASE_DN}
-objectClass: groupOfNames
-cn: developers
-description: Development team
-member: uid=testuser-1,ou=users,${LDAP_BASE_DN}
-EOF
-
-# Add test data to LDAP
-sudo ldapadd -x -D "${LDAP_ADMIN_DN}" -w "${LDAP_ADMIN_PASSWORD}" -f /tmp/test_users.ldif
-
 # Create configuration file for easy reference
 cat << EOF > /tmp/openldap_config.txt
 LDAP Server Configuration:
