+++ deployment/terraform/engine.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"io"
 	"os/exec"
-	"strings"
 
 	"time"
 
@@ -39,7 +38,7 @@ func (t *Terraform) runCommand(dst io.Writer, args ...string) error {
 	mlog.Debug("Running terraform command", mlog.String("args", fmt.Sprintf("%v", args)))
 	cmd := exec.CommandContext(ctx, terraformBin, args...)
 
-	return RunCommand(cmd, dst)
+	return deployment.RunCommand(cmd, dst)
 }
 
 func (t *Terraform) runAWSCommand(ctx context.Context, args []string, dst io.Writer) error {
@@ -61,36 +60,7 @@ func (t *Terraform) runAWSCommand(ctx context.Context, args []string, dst io.Wri
 	mlog.Debug("Running aws command", mlog.String("args", fmt.Sprintf("%v", args)))
 	cmd := exec.CommandContext(ctx, awsBin, args...)
 
-	return RunCommand(cmd, dst)
-}
-
-// CmdLogger implements io.Writer to log command output through mlog
-type CmdLogger struct{}
-
-// Write logs the input to mlog and returns the length of the input
-func (*CmdLogger) Write(in []byte) (int, error) {
-	mlog.Info(strings.TrimSpace(string(in)))
-	return len(in), nil
-}
-
-// RunCommand executes a command with proper logging
-// If dst is set, it captures output to dst. Otherwise, it logs output through mlog.
-func RunCommand(cmd *exec.Cmd, dst io.Writer) error {
-	// If dst is set, that means we want to capture the output.
-	// We write a simple case to handle that using CombinedOutput.
-	if dst != nil {
-		out, err := cmd.CombinedOutput()
-		if err != nil {
-			return err
-		}
-		_, err = dst.Write(out)
-		return err
-	}
-
-	cmd.Stdout = &CmdLogger{}
-	cmd.Stderr = cmd.Stdout
-
-	return cmd.Run()
+	return deployment.RunCommand(cmd, dst)
 }
 
 func checkTerraformVersion() error {
